arc scripts working. cleanup for notebook
