+++ Projects/ABM_DA/experiments/ukf_experiments/ukf_ex2.py
@@ -6,73 +6,18 @@ Created on Mon Dec  2 11:19:48 2019
 @author: rob
 """
 import sys
-sys.path.append("..")
-from ukf_experiments.ukf_fx import fx
+try:
+    sys.path.append("..")
+    from ukf_experiments.ukf_fx import fx
+    from ukf_experiments.poly_functions import poly_count, grid_poly
+    from ukf_experiments.ukf_plots import ukf_plots
+except:
+    sys.path.append("../experiments/ukf_experiments")
+    from ukf_fx import fx
+    from poly_functions import poly_count, grid_poly
+    from ukf_plots import ukf_plots
 
 import numpy as np
-from shapely.geometry import Polygon,MultiPoint
-from shapely.prepared import prep
-
-def poly_count(poly_list,points):
-    
-    
-    """ counts how many agents in each closed polygon of poly_list
-    
-    -    use shapely polygon functions to count agents in each polygon
-    
-    Parameters
-    ------
-    poly_list : list
-        list of closed polygons over StationSim corridor `poly_list`
-    points : array_like
-        list of agent GPS positions to count
-    
-    Returns
-    ------
-    counts : array_like
-        `counts` how many agents in each polygon
-    """
-    counts = []
-    points = np.array([points[::2],points[1::2]]).T
-    points =MultiPoint(points)
-    for poly in poly_list:
-        poly = prep(poly)
-        counts.append(int(len(list(filter(poly.contains,points)))))
-    return counts
-    
-def grid_poly(width,length,bin_size):
-    
-    
-    """generates complete grid of tesselating square polygons covering corridor in station sim.
-   
-    Parameters
-    -----
-    width,length : float
-        `width` and `length` of StationSim corridor. 
-    
-    bin_size : float
-     size of grid squares. larger implies fewer squares `bin_size`
-     
-    Returns
-    ------
-    polys : list
-        list of closed square polygons `polys`
-    """
-    polys = []
-    for i in range(int(width/bin_size)):
-        for j in range(int(length/bin_size)):
-            bl = [x*bin_size for x in (i,j)]
-            br = [x*bin_size for x in (i+1,j)]
-            tl = [x*bin_size for x in (i,j+1)]
-            tr = [x*bin_size for x in (i+1,j+1)]
-            
-            polys.append(Polygon((bl,br,tr,tl)))
-    "hashed lines for plots to verify desired grid"
-    #for poly in polys:
-    #    plt.plot(*poly.exterior.xy)
-    return polys
-
-
 def obs_key_func(state,model_params,ukf_params):
         """which agents are observed"""
         
@@ -140,3 +85,28 @@ def hx2(state,model_params,ukf_params):
         counts = poly_count(ukf_params["poly_list"],state)
         
         return counts
+    
+def ex2_plots(instance,plot_dir,animate,prefix):
+    plts = ukf_plots(instance,plot_dir,prefix)
+        
+    "pull data and put finished agents to nan"
+    obs,preds,full_preds,truth,obs_key,nan_array= instance.data_parser()
+    ukf_params = instance.ukf_params
+    truth[~nan_array]=np.nan
+    preds[~nan_array]=np.nan
+    full_preds[~nan_array]=np.nan
+
+    plts.pair_frame(truth, preds, obs_key, 50)
+    plts.heatmap_frame(truth,ukf_params,50)
+    plts.error_hist(truth, preds, False)
+    plts.path_plots(truth,preds, False)
+    
+    if animate:
+                
+        #plts.trajectories(truth)
+        plts.heatmap(truth,ukf_params,truth.shape[0])
+        if ukf_params["sample_rate"]>1:
+            plts.pair_frames_animation(truth,full_preds,range(truth.shape[0]))
+        else:
+            plts.pair_frames_animation(truth,preds)
+
